requirements.txt
opencv-python~=4.9.0.80
numpy~=1.26.4
insightface~=0.7.3
ultralytics~=8.3.34
redis~=5.2.0
flask~=3.1.0
werkzeug~=3.1.3
pillow~=10.2.0